moved file location of website setup
