+++ main.py
@@ -1,7 +1,7 @@
 from pkg.file_opener import open_file
 from pkg.excel_parser import parse_excel
-from pkg.setup_website import setup
-from pkg.fill_users import fill_users
+from pkg.setup_website.setup_website import setup
+from pkg.setup_website.fill_users import fill_users
 
 
 def main():
